Changes to POA_Website
Created api package to contain all non frontend routes, this means if it changes something in the database or supplies a modal we
don't want it here this gives us an advantage in the future if we want to integrate this app with other web apps

Moved from Main to api
	popUpMessage(title, message)
	addTrip()
	addParticipant(FormKey)
	editParticipant(FormKey)
	checkAddParticipant(FormKey)
	cannotLeave() -> should be depricated soon change to use popUpMessage and Javascript
	removeParticipant(personID, tripID)
	swapCoordinators(oldLeaderID, newLeaderID, tripID)

Just came to me you can test with multipe users if u use two google accounts in two diffrent browsers at the same time

Moved the contence of swapCoordinators to DatabaseQuery as a query for Particpants

Created controllerHelperMethods
	tripPage -> move fill bar calculations to calculateProgress_carRatio(masterInfo), and calculateProgress_participantRatio(masterInfo)
		if you could add doc strings here Im not completly clear on how the calculation works

making changes to how account is created changing constructor to take formData and session so that the constructor takes in the raw data and parses it eliminating a layer of data processing to construct an account.\
Removing id feild as the users goog num this doesn't need to be in two feilds and could potentially be cause problems though prob not. I also added a unique paraperter to gooogle num since we are assuming these numbers are unique and we don't want people to be able to make multiple accounts

We should make a custom validator for student id nums

Your comment: Also given a picture. That's neat. How to access it?
Awnser: we are give a url to where a pic is hosted by google we can show it by using that url and a image tag

Considering making create account a modal

moved create account into Account class and added arguments so it serves as a constructor that commits to the db

moving modify account to database query in class Participant_manipulation_query

can you add coments to 186-194 Main/controllers.py I found this code confusing

I'd do a quick test of all functionality I was unable to test anything requiring login but I am confident that nothing should be broken the most likley bug is that something needs to be would be a redirect from main.NAME to api.NAME